forwardport of rev 4021 from 1.0 branch
git-svn-id: svn://svn.camaya.net/gloox/trunk@4022 bb60cc1b-8fe5-0310-9879-aca5d778a633

+++ src/sha.cpp
@@ -128,7 +128,7 @@ namespace gloox
 
   void SHA::feed( const std::string& data )
   {
-    feed( (const unsigned char*)data.c_str(), data.length() );
+    feed( (const unsigned char*)data.c_str(), (int)data.length() );
   }
 
   void SHA::process()
